+++ src/Illusion/Graphics/PipelineResource.hpp
@@ -78,7 +78,7 @@ struct PipelineResource {
   uint32_t             mColumns              = 0;
   uint32_t             mArraySize            = 0;
   uint32_t             mOffset               = 0;
-  uint32_t             mSize                 = 0;
+  size_t               mSize                 = 0;
   std::string          mName;
   std::vector<Member>  mMembers;
 };
